# Conflicts:
#	FleetManager/Models/Vehicle.cs

+++ FleetManager/Models/Vehicle.cs
@@ -7,6 +7,6 @@ public class Vehicle : ReactiveObject
     [Reactive] public int VehicleId { get; set; }
     [Reactive] public string VehicleName { get; set; } = string.Empty;
     [Reactive] public string VehicleTag { get; set; } = string.Empty;
-    [Reactive] public int VehicleFuel { get; set; }
+    [Reactive] public double VehicleFuel { get; set; }
     [Reactive] public string VehicleStatus { get; set; } = string.Empty;
 }
